vfs: run Release asynchronously for read write files handles
Before this change the vfs layer would run Release asynchronously.
This would mean that it would wait for cached files to be uploaded
before returning.

However Release is already running asynchronously from userspace so
this commit changes it to do the functionality of Release
asynchronously too.

This should fix libfuse blocking when Release is active and it is
asked to do something else with a file.

Forum: https://forum.rclone.org/t/vfs-cache-mode-writes-upload-expected-behaviour/8014